web_app/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pandas==2.3.1

llama-index-core==0.11.0
llama_index-vector_stores-faiss==0.2.1
llama-index-llms-openai==0.2.0
llama-index-llms-azure-openai==0.2.0
llama-index-llms-ollama==0.3.0
llama-index-readers-web==0.2.0
llama-index-embeddings-huggingface==0.3.1
openai==1.63.2
faiss-cpu==1.10.0
pymupdf4llm==0.0.17
sentence_transformers==3.4.1
httpx==0.28.1
mammoth==1.9.0
markdownify==0.13.1
pyyaml==6.0.1
tqdm==4.66.5
beautifulsoup4==4.13.3
playwright==1.46.0
einops==0.8.1
doc2text==0.2.4
marker-pdf==1.6.1
FlagEmbedding==1.3.4
jupyterlab==4.4.2
docx2txt==0.9
transformers==4.45.2
tabulate==0.9.0
ipywidgets==8.1.7
